Implement initial reporting of rule specification total coverages

+++ core/core/job.py
@@ -59,6 +59,7 @@ class Job(core.utils.CallbacksCaller):
         self.uploading_reports_process_exitcode = None
         self.data = None
         self.data_lock = None
+        self.rule_spec_total_coverage_files_and_arcnames = None
         self.type = type
         self.components_common_conf = None
         self.sub_jobs = []
@@ -66,6 +67,7 @@ class Job(core.utils.CallbacksCaller):
         self.callbacks = {}
         self.component_processes = []
         self.reporting_results_process = None
+        self.collecting_rule_spec_total_coverages_process = None
 
     def decide(self, conf, mqs, locks, vals, uploading_reports_process_exitcode):
         self.logger.info('Decide job')
@@ -216,24 +218,27 @@ class Job(core.utils.CallbacksCaller):
                     # TODO: if self.components_common_conf['VTG strategy']['collect total coverage'] != 'None':
                     if True:
                         context.logger.info('Terminate rule specification coverage infos message queue')
-                        context.mqs['rule specification coverage infos'].put(None)
+                        context.mqs['rule specification coverage info files'].put(None)
 
                 core.utils.set_component_callbacks(self.logger, type(self), (after_finish_task_results_processing,))
 
                 # TODO: if self.components_common_conf['VTG strategy']['collect total coverage'] != 'None':
                 if True:
-                    self.mqs['rule specification coverage infos'] = multiprocessing.Queue()
+                    self.mqs['rule specification coverage info files'] = multiprocessing.Queue()
 
                     def after_process_finished_task(context):
-                        context.mqs['rule specification coverage infos'].put({
-                            'rule specification': context.rule_specification,
-                            'coverage info': context.coverage.coverage_info
-                        })
+                        if os.path.isfile('coverage info.json'):
+                            context.mqs['rule specification coverage info files'].put({
+                                'rule specification': context.rule_specification,
+                                'coverage info file': os.path.relpath('coverage info.json',
+                                                                      context.conf['main working directory'])
+                            })
 
                     core.utils.set_component_callbacks(self.logger, type(self), (after_process_finished_task,))
 
-                    self.collect_coverage_process = multiprocessing.Process(target=self.collect_coverage)
-                    self.collect_coverage_process.start()
+                    self.collecting_rule_spec_total_coverages_process = \
+                        multiprocessing.Process(target=self.collect_rule_spec_total_coverage)
+                    self.collecting_rule_spec_total_coverages_process.start()
 
                 self.get_sub_job_components()
 
@@ -270,15 +275,23 @@ class Job(core.utils.CallbacksCaller):
                     core.utils.remove_component_callbacks(self.logger, type(self))
 
                     if self.name:
-                        core.utils.report(self.logger,
-                                          'finish',
-                                          {
-                                              'id': self.id,
-                                              'resources': {'wall time': 0, 'CPU time': 0, 'memory size': 0},
-                                          },
-                                          self.mqs['report files'],
-                                          self.vals['report id'],
-                                          self.components_common_conf['main working directory'])
+                        report = {
+                            'id': self.id,
+                            'resources': {'wall time': 0, 'CPU time': 0, 'memory size': 0},
+                        }
+
+                        if len(self.rule_spec_total_coverage_files_and_arcnames):
+                            report['coverage'] = {}
+
+                            for rule_spec, total_coverage_files_and_arcnames in \
+                                    self.rule_spec_total_coverage_files_and_arcnames.items():
+                                total_coverage_file = total_coverage_files_and_arcnames['total coverage file']
+                                arcnames = total_coverage_files_and_arcnames['arcnames']
+                                report['coverage'][rule_spec] = core.utils.ReportFiles(
+                                    [total_coverage_file] + list(arcnames.keys()), arcnames=arcnames)
+
+                        core.utils.report(self.logger, 'finish', report, self.mqs['report files'],
+                                          self.vals['report id'], self.components_common_conf['main working directory'])
                 except Exception:
                     self.logger.exception('Catch exception')
 
@@ -376,6 +389,7 @@ class Job(core.utils.CallbacksCaller):
                 sub_job.uploading_reports_process_exitcode = self.uploading_reports_process_exitcode
                 sub_job.data = self.data
                 sub_job.data_lock = self.data_lock
+                sub_job.rule_spec_total_coverage_files_and_arcnames = multiprocessing.Manager().dict()
                 sub_job.components_common_conf = sub_job_concrete_conf
 
     def get_sub_job_components(self):
@@ -424,6 +438,10 @@ class Job(core.utils.CallbacksCaller):
 
                 if self.reporting_results_process and self.reporting_results_process.exitcode:
                     raise RuntimeError('Reporting results failed')
+
+                if self.collecting_rule_spec_total_coverages_process and \
+                        self.collecting_rule_spec_total_coverages_process.exitcode:
+                    raise RuntimeError('Collecting rule specification total coverages failed')
         except Exception:
             for p in self.component_processes:
                 # Do not terminate components that already exitted.
@@ -434,9 +452,9 @@ class Job(core.utils.CallbacksCaller):
                 self.logger.info('Forcibly terminate verification statuses message queue')
                 self.mqs['verification statuses'].put(None)
 
-            if 'rule specification coverage infos' in self.mqs:
-                self.logger.info('Forcibly terminate rule specification coverage infos message queue')
-                self.mqs['rule specification coverage infos'].put(None)
+            if 'rule specification coverage info files' in self.mqs:
+                self.logger.info('Forcibly terminate rule specification coverage info files message queue')
+                self.mqs['rule specification coverage info files'].put(None)
 
             raise
         finally:
@@ -446,39 +464,72 @@ class Job(core.utils.CallbacksCaller):
                 if self.reporting_results_process.exitcode:
                     raise RuntimeError('Reporting results failed')
 
-    def collect_coverage(self):
-        total_coverage = {}
+            if self.collecting_rule_spec_total_coverages_process:
+                self.logger.info('Wait for collecting all rule specification total coverages')
+                self.collecting_rule_spec_total_coverages_process.join()
+                if self.collecting_rule_spec_total_coverages_process.exitcode:
+                    raise RuntimeError('Collecting rule specification total coverages failed')
 
-        while True:
-            rule_spec_and_coverage_infos = self.mqs['rule specification coverage infos'].get()
-
-            if rule_spec_and_coverage_infos is None:
-                self.logger.debug('Rule specification coverage infos message queue was terminated')
-                self.mqs['rule specification coverage infos'].close()
-                break
-
-            rule_spec = rule_spec_and_coverage_infos['rule specification']
-
-            for file_name, coverage_info in rule_spec_and_coverage_infos['coverage info'].items():
-                total_coverage.setdefault(rule_spec, {})
-                total_coverage[rule_spec].setdefault(file_name, [])
-                total_coverage[rule_spec][file_name] += coverage_info
-
-        for rule, coverage_info in total_coverage.items():
-            # TODO: self.components_common_conf['VTG strategy']['collect total coverage'])
-            coverage = core.vrp.coverage_parser.LCOV.get_coverage(coverage_info, 'full')
-
-            with open('coverage.json', 'w', encoding='utf-8') as fp:
-                json.dump(coverage, fp, ensure_ascii=True, sort_keys=True, indent=4)
-
-            arcnames = {info[0]['file name']: info[0]['arcname'] for info in coverage_info.values()}
-            # TODO: report
-
-    def report_results(self):
-        os.mkdir('results')
-
+    def collect_rule_spec_total_coverage(self):
         # Process exceptions like for uploading reports.
         try:
+            total_coverage_infos = {}
+
+            while True:
+                rule_spec_and_coverage_info_files = self.mqs['rule specification coverage info files'].get()
+
+                if rule_spec_and_coverage_info_files is None:
+                    self.logger.debug('Rule specification coverage info files message queue was terminated')
+                    self.mqs['rule specification coverage info files'].close()
+                    break
+
+                rule_spec = rule_spec_and_coverage_info_files['rule specification']
+                total_coverage_infos.setdefault(rule_spec, {})
+
+                with open(os.path.join(self.components_common_conf['main working directory'],
+                                       rule_spec_and_coverage_info_files['coverage info file']), encoding='utf8') as fp:
+                    coverage_info = json.load(fp)
+
+                for file_name, coverage_info in coverage_info.items():
+                    total_coverage_infos[rule_spec].setdefault(file_name, [])
+                    total_coverage_infos[rule_spec][file_name] += coverage_info
+
+            os.mkdir('total coverages')
+
+            rule_spec_total_coverage_files_and_arcnames = {}
+
+            for rule_spec, coverage_info in total_coverage_infos.items():
+                total_coverage_dir = os.path.join('total coverages', re.sub(r'/', '-', rule_spec))
+                os.mkdir(total_coverage_dir)
+
+                total_coverage_file = os.path.join(total_coverage_dir, 'coverage.json')
+                if os.path.isfile(total_coverage_file):
+                    raise FileExistsError('Rule specification total coverage file "{0}" already exists'
+                                          .format(total_coverage_file))
+
+                # TODO: self.components_common_conf['VTG strategy']['collect total coverage'])
+                coverage = core.vrp.coverage_parser.LCOV.get_coverage(coverage_info, 'full')
+
+                with open(total_coverage_file, 'w', encoding='utf8') as fp:
+                    json.dump(coverage, fp, ensure_ascii=True, sort_keys=True, indent=4)
+
+                rule_spec_total_coverage_files_and_arcnames[rule_spec] = {
+                    'total coverage file': total_coverage_file,
+                    'arcnames': {info[0]['file name']: info[0]['arcname'] for info in coverage_info.values()}
+                }
+
+            # Share collected rule specification total coverages and arcnames to report them within Sub-job finish
+            # report.
+            self.rule_spec_total_coverage_files_and_arcnames.update(rule_spec_total_coverage_files_and_arcnames)
+        except Exception:
+            self.logger.exception('Catch exception when collecting rule specification total coverages')
+            os._exit(1)
+
+    def report_results(self):
+        # Process exceptions like for uploading reports.
+        try:
+            os.mkdir('results')
+
             while True:
                 verification_status = self.mqs['verification statuses'].get()
 
@@ -517,7 +568,7 @@ class Job(core.utils.CallbacksCaller):
                                       self.vals['report id'],
                                       self.components_common_conf['main working directory'],
                                       results_dir)
-        except Exception as e:
+        except Exception:
             self.logger.exception('Catch exception when reporting results')
             os._exit(1)
 
@@ -577,8 +628,7 @@ class Job(core.utils.CallbacksCaller):
 
     def __process_validation_results(self, name, verification_result):
         # Relate verificaiton results on commits before and after corresponding bug fixes if so.
-        # Without this we won't be able to reliably iterate over data since it is
-        # multiprocessing.Manager().dict().
+        # Without this we won't be able to reliably iterate over data since it is multiprocessing.Manager().dict().
         # Data is intended to keep verification results that weren't bound still. For such the results
         # we will need to update corresponding data sent before.
         data = self.data.copy()
